Fix jshint
